Rewrite rules for bind (#157)
This commit defines `fmap` in terms of `bind` and adds a test to make sure that `fmap f . fmap g` is rewritten to `fmap (f.g)` by the rewrite rules.

This change reflects that `fmap` can be expressed with `bind` and `bind` can be fused with `foldg`. Moreover, any function preserving the structure of the graph (like `fmap`) can be expressed with `bind`, so this is an "optimal" change, since we do not have to add any other rule for structure preserving-functions.