Added functionality to the 'SymmetricRelation' module. Functionalites (based on 'Algebra.Graph.Relation'):     - Basic graph construction primitives     - Relations on graphs     - Graph properties     - Standard families of graphs     - Graph transformation     - Relational operations (only 'compose' yet) These new functions respect the undirected graph axiom (commutativity of 'connect').
Adjusted the implementation of the 'SymmetricRelation' 'Graph' instance
to respect the axiom.

Problems: 'Num' instance for 'SymmetricRelation' was derived so the
axiom was not guaranteed for (*), so I added a custom instance
implementation on 'Algebra.Graph.Relation.Symmetric' which the compiler
gives an error for duplicate declaration of the 'Num' instance. I
already deleted the 'deriving Num' from the 'Relation.InternalDerived'
module but the compiler still gives the error.